Update README.md
Hi, your video was one of the best instructional videos I've ever seen on any robotics/electronics project, great work. We did build this board for my mother with PD. She has been using it 2 days without relief, but we are hopeful it will help. We are going to build the vibration sensors next to ensure it is working correctly, and then work on the gloves so she can have more mobility. In this docs I included a video of it in operation for her in an early state of assembly. My hope is that my next build I can add much more instructions, and I have some additional photos and videos I want to include in an upcoming readme update. We know others with PD who are interested in assembling their own, so I would like to continue to document everything for my own use helping others. If you like it, I'd love your repo to be the source of truth for this, since this is really your work. 

However, I'm not offended by you wanting to make changes or reject these entirely. Thanks again for everything!